Extracted the SearchMovies component logic into a new useSearchMovies custom hook
